requirements.txt
asgiref==3.7.2
Django==3.1.5
djangorestframework==3.14.0
sqlparse==0.4.4
